requirements.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
django-filter==24.2
drf-yasg==1.21.7
iniconfig==2.1.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.2
sqlparse==0.5.3
